Add comprehensive benchmarks and edge case tests for Orus language
- Introduced extreme benchmarks in `extreme_benchmark.py` to stress test arithmetic, variable pressure, expression complexity, register allocation, and integration.
- Created a unified benchmark script `unified_benchmark.sh` for cross-language performance comparison.
- Added various edge case tests in `arithmetic_edge_cases.orus`, `boundary_values.orus`, `error_conditions.orus`, `expression_nesting.orus`, `literal_edge_cases.orus`, `operator_precedence.orus`, `print_edge_cases.orus`, and `variable_edge_cases.orus` to validate the robustness of the Orus language.
- Each edge case file tests specific scenarios including zero operations, large numbers, deep nesting, operator precedence, and variable handling.

+++ include/builtins.h
@@ -9,7 +9,7 @@
 void builtin_print(Value* args, int count, bool newline);
 
 // High-precision timestamp function
-// Returns nanoseconds since an arbitrary but monotonic starting point
-int64_t builtin_time_stamp(void);
+// Returns milliseconds since an arbitrary but monotonic starting point
+int32_t builtin_time_stamp(void);
 
 #endif // ORUS_BUILTINS_H
